Created Boilerplate For Recipe Pages
I have completed work on the homepage, including the necessary information and links. Additionally, the Spaghetti Carbonara page is now finished,
featuring a list of ingredients and a step by step guide on how to make the dish.
The next step is to focus on completing the remaining pages for Lasagna, Pizza, and Vongole.